+++ lib/in_app_console.dart
@@ -5,3 +5,4 @@ export 'package:in_app_console/src/core/logger/in_app_logger_data.dart';
 export 'package:in_app_console/src/core/logger/in_app_logger_type.dart';
 export 'package:in_app_console/src/core/console/in_app_console.dart';
 export 'package:in_app_console/src/core/extension/in_app_console_extension.dart';
+export 'package:in_app_console/src/core/extension/in_app_console_extension_context.dart';