+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 2.0.0
+
+### Breaking Changes
+* **BREAKING:** Renamed `clearHistory()` to `clearLogs()` - Update your code to use the new method name
+
+### New Features
+* Added extension system for custom functionality via `InAppConsoleExtension`
+* Added `registerExtension()` and `unregisterExtension()` methods
+* Added `InAppConsoleExtensionContext` for extension data access
+
+### Improvements
+* UI improvements and bug fixes
+
 ## 1.1.0
 * Added `InAppConsole.kEnableConsole` flag
 * Changed `error` property data type: